requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.1
rsa==4.9
scipy==1.14.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.14.1
